Remove `table_slice_builder`
