+++ libtenzir/builtins/aspects/connectors.cpp
@@ -9,25 +9,13 @@
 #include <tenzir/argument_parser.hpp>
 #include <tenzir/collect.hpp>
 #include <tenzir/plugin.hpp>
-#include <tenzir/table_slice_builder.hpp>
+#include <tenzir/series_builder.hpp>
 
 #include <algorithm>
 
 namespace tenzir::plugins::connectors {
 
 namespace {
-
-/// A type that represents a connector.
-auto connector_type() -> type {
-  return type{
-    "tenzir.connector",
-    record_type{
-      {"name", string_type{}},
-      {"loader", bool_type{}},
-      {"saver", bool_type{}},
-    },
-  };
-}
 
 class plugin final : public virtual aspect_plugin {
 public:
@@ -37,25 +25,33 @@ public:
 
   auto show(operator_control_plane& ctrl) const
     -> generator<table_slice> override {
+    TENZIR_UNUSED(ctrl);
     auto loaders = collect(plugins::get<loader_parser_plugin>());
     auto savers = collect(plugins::get<saver_parser_plugin>());
     auto connectors = std::set<std::string>{};
-    for (const auto* plugin : loaders)
+    for (const auto* plugin : loaders) {
       connectors.insert(plugin->name());
-    for (const auto* plugin : savers)
-      connectors.insert(plugin->name());
-    auto builder = table_slice_builder{connector_type()};
-    for (const auto& connector : connectors) {
-      if (not(builder.add(connector)
-              && builder.add(plugins::find<loader_parser_plugin>(connector)
-                             != nullptr)
-              && builder.add(plugins::find<saver_parser_plugin>(connector)
-                             != nullptr))) {
-        diagnostic::error("failed to add connector").emit(ctrl.diagnostics());
-        co_return;
-      }
     }
-    co_yield builder.finish();
+    for (const auto* plugin : savers) {
+      connectors.insert(plugin->name());
+    }
+    auto builder = series_builder{type{
+      "tenzir.connector",
+      record_type{
+        {"name", string_type{}},
+        {"loader", bool_type{}},
+        {"saver", bool_type{}},
+      },
+    }};
+    for (const auto& connector : connectors) {
+      auto event = builder.record();
+      event.field("name", connector);
+      event.field("loader",
+                  plugins::find<loader_parser_plugin>(connector) != nullptr);
+      event.field("saver",
+                  plugins::find<saver_parser_plugin>(connector) != nullptr);
+    }
+    co_yield builder.finish_assert_one_slice();
   }
 };
