+++ libtenzir/builtins/contexts/lookup_table.cpp
@@ -25,7 +25,6 @@
 #include <tenzir/series_builder.hpp>
 #include <tenzir/session.hpp>
 #include <tenzir/table_slice.hpp>
-#include <tenzir/table_slice_builder.hpp>
 #include <tenzir/tql2/eval.hpp>
 #include <tenzir/tql2/plugin.hpp>
 #include <tenzir/type.hpp>
