+++ libtenzir/builtins/functions/hash.cpp
@@ -25,7 +25,6 @@
 #include <tenzir/hash/xxhash.hpp>
 #include <tenzir/optional.hpp>
 #include <tenzir/plugin.hpp>
-#include <tenzir/table_slice_builder.hpp>
 #include <tenzir/tql2/plugin.hpp>
 
 #include <arrow/scalar.h>
