+++ libtenzir/builtins/operators/unroll.cpp
@@ -13,7 +13,6 @@
 #include <tenzir/collect.hpp>
 #include <tenzir/fwd.hpp>
 #include <tenzir/plugin.hpp>
-#include <tenzir/table_slice_builder.hpp>
 #include <tenzir/tql2/ast.hpp>
 #include <tenzir/tql2/eval.hpp>
 #include <tenzir/tql2/plugin.hpp>
