+++ libtenzir/src/tql2/eval_impl.cpp
@@ -6,6 +6,7 @@
 // SPDX-FileCopyrightText: (c) 2024 The Tenzir Contributors
 // SPDX-License-Identifier: BSD-3-Clause
 
+#include <tenzir/arrow_table_slice.hpp>
 #include <tenzir/arrow_utils.hpp>
 #include <tenzir/checked_math.hpp>
 #include <tenzir/collect.hpp>
@@ -13,7 +14,6 @@
 #include <tenzir/detail/zip_iterator.hpp>
 #include <tenzir/multi_series_builder.hpp>
 #include <tenzir/series_builder.hpp>
-#include <tenzir/table_slice_builder.hpp>
 #include <tenzir/tql2/eval_impl.hpp>
 
 #include <ranges>
