+++ nix/tenzir/plugins/source.json
@@ -2,7 +2,8 @@
   "name": "tenzir-plugins",
   "url": "git@github.com:tenzir/tenzir-plugins",
   "ref": "main",
-  "rev": "63e91ffc0625a570adc4c2f56c7f2f90ac7eda2f",
+  "rev": "5359067b207b9e578192b7f77c658a1fda737b09",
   "submodules": true,
-  "shallow": true
+  "shallow": true,
+  "allRefs": true
 }
