+++ plugins/nic/builtins/plugin.cpp
@@ -17,7 +17,6 @@
 #include <tenzir/logger.hpp>
 #include <tenzir/pcap.hpp>
 #include <tenzir/plugin.hpp>
-#include <tenzir/table_slice_builder.hpp>
 #include <tenzir/tql2/plugin.hpp>
 
 #include <pcap/pcap.h>
